fix(*): Fix duplicated fetch
