--- src/components/__FolderNameList/FolderNameList.module.css
@@ -1,20 +1,0 @@
-/* .folder-buttons-section {
-  padding: 2rem 15rem 2rem 4rem;
-}
-
-.folder-buttons-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
-
-.folder-title {
-  color: #000;
-  margin-top: 5rem;
-  font-family: Pretendard;
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  letter-spacing: -0.2px;
-} */
